+++ sketch-view-core/src/androidTest/kotlin/com/github/panpf/sketch/view/core/test/drawable/ResizeDrawableTest.kt
@@ -51,6 +51,8 @@ import kotlin.test.assertTrue
 @RunWith(AndroidJUnit4::class)
 class ResizeDrawableTest {
 
+    // TODO test scaleType
+
     @Test
     fun testResize() {
         ColorDrawable(Color.GREEN).resize(Size(100, 100)).apply {