+++ sketch-view-core/src/main/kotlin/com/github/panpf/sketch/drawable/ResizeDrawable.kt
@@ -18,20 +18,35 @@ package com.github.panpf.sketch.drawable
 
 import android.graphics.drawable.Animatable
 import android.graphics.drawable.Drawable
+import android.widget.ImageView.ScaleType
 import androidx.appcompat.graphics.drawable.DrawableWrapperCompat
 import com.github.panpf.sketch.resize.Scale
-import com.github.panpf.sketch.resize.Scale.CENTER_CROP
 import com.github.panpf.sketch.util.Size
-import com.github.panpf.sketch.util.calculateBounds
-import com.github.panpf.sketch.util.toAndroidRect
+import com.github.panpf.sketch.util.calculateBoundsWithScaleType
 import com.github.panpf.sketch.util.toLogString
+import com.github.panpf.sketch.util.toScale
+import com.github.panpf.sketch.util.toScaleType
 
 /**
  * Using [size] as the intrinsic size of [drawable], [drawable] will be scaled according to the scale of [size].
  *
  * @see com.github.panpf.sketch.view.core.test.drawable.ResizeDrawableTest.testResize
  */
-fun Drawable.resize(size: Size, scale: Scale = CENTER_CROP): ResizeDrawable {
+fun Drawable.resize(size: Size, scaleType: ScaleType = ScaleType.CENTER_CROP): ResizeDrawable {
+    return if (this is Animatable) {
+        ResizeAnimatableDrawable(this, size, scaleType)
+    } else {
+        ResizeDrawable(this, size, scaleType)
+    }
+}
+
+/**
+ * Using [size] as the intrinsic size of [drawable], [drawable] will be scaled according to the scale of [size].
+ *
+ * @see com.github.panpf.sketch.view.core.test.drawable.ResizeDrawableTest.testResize
+ */
+@Deprecated("Use resize(size, scaleType) instead")
+fun Drawable.resize(size: Size, scale: Scale): ResizeDrawable {
     return if (this is Animatable) {
         ResizeAnimatableDrawable(this, size, scale)
     } else {
@@ -45,12 +60,25 @@ fun Drawable.resize(size: Size, scale: Scale = CENTER_CROP): ResizeDrawable {
  *
  * @see com.github.panpf.sketch.view.core.test.drawable.ResizeDrawableTest
  */
-// TODO Support ScaleType
-open class ResizeDrawable(
+open class ResizeDrawable constructor(
     drawable: Drawable,
     val size: Size,
-    val scale: Scale
+    val scaleType: ScaleType = ScaleType.CENTER_CROP
 ) : DrawableWrapperCompat(drawable), SketchDrawable {
+
+    @Deprecated("Use scaleType instead.")
+    val scale: Scale = scaleType.toScale()
+
+    @Deprecated(message = "Use ResizeDrawable(painter, size, scaleType) instead")
+    constructor(
+        drawable: Drawable,
+        size: Size,
+        scale: Scale
+    ) : this(
+        drawable = drawable,
+        size = size,
+        scaleType = scale.toScaleType()
+    )
 
     override fun getIntrinsicWidth(): Int {
         return size.width
@@ -62,25 +90,33 @@ open class ResizeDrawable(
 
     override fun setBounds(left: Int, top: Int, right: Int, bottom: Int) {
         super.setBounds(left, top, right, bottom)
-        drawable?.apply {
-            this@apply.bounds = calculateBounds(
-                srcSize = Size(
-                    width = this@apply.intrinsicWidth,
-                    height = this@apply.intrinsicHeight
-                ),
-                dstSize = Size(
-                    width = this@ResizeDrawable.bounds.width(),
-                    height = this@ResizeDrawable.bounds.height()
-                ),
-                scale = scale
-            ).toAndroidRect()
+        val drawable = drawable ?: return
+        val srcSize = Size(
+            width = drawable.intrinsicWidth,
+            height = drawable.intrinsicHeight
+        )
+        if (srcSize.isEmpty) {
+            drawable.setBounds(left, top, right, bottom)
+            return
         }
+
+        val drawableBounds = calculateBoundsWithScaleType(
+            srcSize = srcSize,
+            dstSize = Size(width = right - left, height = bottom - top),
+            scaleType = scaleType
+        )
+        drawable.setBounds(
+            /* left = */ drawableBounds.left,
+            /* top = */ drawableBounds.top,
+            /* right = */ drawableBounds.right,
+            /* bottom = */ drawableBounds.bottom
+        )
     }
 
     override fun mutate(): ResizeDrawable {
         val mutateDrawable = drawable?.mutate()
         return if (mutateDrawable != null && mutateDrawable !== drawable) {
-            ResizeDrawable(mutateDrawable, size, scale)
+            ResizeDrawable(mutateDrawable, size, scaleType)
         } else {
             this
         }
@@ -92,17 +128,17 @@ open class ResizeDrawable(
         other as ResizeDrawable
         if (size != other.size) return false
         if (drawable != other.drawable) return false
-        return scale == other.scale
+        return scaleType == other.scaleType
     }
 
     override fun hashCode(): Int {
         var result = size.hashCode()
-        result = 31 * result + scale.hashCode()
+        result = 31 * result + scaleType.hashCode()
         result = 31 * result + drawable.hashCode()
         return result
     }
 
     override fun toString(): String {
-        return "ResizeDrawable(drawable=${drawable?.toLogString()}, size=$size, scale=$scale)"
+        return "ResizeDrawable(drawable=${drawable?.toLogString()}, size=$size, scaleType=$scaleType)"
     }
 }